Basic support for prepared statements on sqlite3.
With this modification we can enjoy the security benefits of having
prepared-statement-alike additional parameters.  To do this, the
additional parameters should be passed after the statement in the
execute method.

This means that a new prepared statement will be created on each
execute call, so don't expect big a performance increase.

Maybe in a distant future a LRU cache of prepared statements could be
added.