ls_sqlite3: conditional use of lua_isinteger
The `lua_isinteger` function is only present in Lua 5.3. To maintain
compatibility with earlier Lua versions (notably 5.2 and 5.1) this
commit gates the usage of this function with `defined` pre-processor
checks. In the case where the symbol is not defined, we unconditionally
fall back to existing logic that treats the maybe-an-int item in
question as if `lua_isinteger` returned false.