Implement dates and timestamps in OCI driver
It is now possible to directly query DATE and TIMESTAMP data types.

There is no direct support for these types in the original OCI interface
version 8, so the code is guarded by ifdef checks.

Unfortunately, there is no interface to query the buffer size required
for output strings. In order not to overcomplicate the code, a
conservative size of 65 bytes (including NUL) was chosen. The "fsprec"
parameter is not specified clearly, so a value of "6" was chosen for
microsecond precision.